+++ R/ao.R
@@ -1,299 +1,336 @@
 #' Alternating Optimization
 #'
 #' @description
-#' This function performs alternating optimization of the function \code{f}.
+#' Alternating optimization is an iterative procedure for optimizing a
+#' real-valued function jointly over all its parameters by alternating
+#' restricted optimization over parameter partitions.
 #'
-#' @param f
-#' A \code{function} to be optimized, returning a single \code{numeric}.
-#' The first argument of \code{f} must be a \code{numeric} of the same length as
-#' \code{p} followed by any other arguments specified by the \code{...}
-#' argument.
-#' @param p
-#' A \code{numeric} vector, the starting parameter values for the optimization.
+#' @param f (`function`)\cr
+#' A \code{function} to be optimized, returning a single \code{numeric} value.
+#'
+#' The first argument of \code{f} should be a \code{numeric} of the same length
+#' as \code{initial}, followed by any other arguments specified by the
+#' \code{...} argument.
+#'
+#' If \code{f} is to be optimized over an argument other than the first, or more
+#' than one argument, this has to be specified via the \code{target} argument.
+#'
+#' @param initial (`numeric()`)\cr
+#' The starting parameter values for the target argument(s).
+#'
+#' @param target (`character()` or `NULL`)\cr
+#' The name(s) of the argument(s) over which \code{f} gets optimized.
+#'
+#' This can only be \code{numeric} arguments.
+#'
+#' Can be `NULL` (default), then it is the first argument of `f`.
+#'
+#' @param npar (`integer()`)\cr
+#' The length of the target argument(s).
+#'
+#' Must be specified if more than two target arguments are specified via
+#' the `target` argument.
+#'
+#' Can be `NULL` if there is only one target argument, in which case `npar` is
+#' set to be `length(initial)`.
+#'
+#' @param gradient (`function` or `NULL`)\cr
+#' A \code{function} that returns the gradient of \code{f}.
+#'
+#' The function call of \code{gradient} must be identical to \code{f}.
+#'
+#' Can be `NULL`, in which case a finite-difference approximation will be used.
+#'
 #' @param ...
-#' Additional arguments to be passed to \code{f}.
-#' @param partition
-#' A \code{list} of vectors of indices of \code{p}, specifying the partition of
-#' the parameter vector in the alternating optimization process.
-#' The default is \code{as.list(1:length(p))}, i.e. each parameter is
-#' optimized separately.
-#' Parameter indices can be members of multiple subsets.
-#' @param base_optimizer
+#' Additional arguments to be passed to \code{f} (and \code{gradient}).
+#'
+#' @param partition (`character(1)` or `list()`)\cr
+#' Defines the parameter partition, and can be either
+#' * `"sequential"` for treating each parameter separately,
+#' * `"random"` for a random partition in each iteration,
+#' * `"none"` for no partition (which is equivalent to joint optimization),
+#' * or a `list` of vectors of parameter indices, specifying a custom
+#'   partition for the alternating optimization process.
+#'
+#' @param new_block_probability (`numeric(1)`)\cr
+#' Only relevant if `partition = "random"`.
+#' The probability for a new parameter block when creating a random
+#' partitions.
+#' Values close to 0 result in larger parameter blocks, values close to 1
+#' result in smaller parameter blocks.
+#'
+#' @param minimum_block_number (`integer(1)`)\cr
+#' Only relevant if `partition = "random"`.
+#' The minimum number of blocks in random partitions.
+#'
+#' @param minimize (`logical(1)`)\cr
+#' Whether to minimize during the alternating optimization process.
+#' If \code{FALSE}, maximization is performed.
+#'
+#' @param lower (`numeric()`)\cr
+#' Optionally lower parameter bounds.
+#'
+#' @param upper (`numeric()`)\cr
+#' Optionally upper parameter bounds.
+#'
+#' @param iteration_limit (`integer(1)` or `Inf`)\cr
+#' The maximum number of iterations through the parameter partition before
+#' the alternating optimization process is terminated.
+#' Can also be `Inf` for no iteration limit.
+#'
+#' @param seconds_limit (`numeric(1)`)\cr
+#' The time limit in seconds before the alternating optimization process is
+#' terminated.
+#' Can also be `Inf` for no time limit.
+#' Note that this stopping criteria is only checked *after*
+#' a sub-problem is solved and not *within* solving a sub-problem, so the
+#' actual process time can exceed this limit.
+#'
+#' @param tolerance_value (`numeric(1)`)\cr
+#' A non-negative tolerance value. The alternating optimization terminates
+#' if the absolute difference between the current function value and the one
+#' from the last iteration is smaller than \code{tolerance_value}.
+#' Can be `0` for no value threshold.
+#'
+#' @param tolerance_parameter (`numeric(1)`)\cr
+#' A non-negative tolerance value. The alternating optimization terminates
+#' if the distance between the current estimate and the one from the last
+#' iteration is smaller than \code{tolerance_parameter}.
+#' Can be `0` for no parameter threshold.
+#' By default, the distance is measured using the euclidean norm, but
+#' another norm can be specified via the \code{tolerance_parameter_norm}
+#' field.
+#'
+#' @param tolerance_parameter_norm (`function`)\cr
+#' The norm that measures the distance between the current estimate and the
+#' one from the last iteration. If the distance is smaller than
+#' \code{tolerance_parameter}, the procedure is terminated.
+#' It must be of the form \code{function(x, y)} for two vector inputs
+#' \code{x} and \code{y}, and return a single \code{numeric} value.
+#' By default, the euclidean norm \code{function(x, y) sqrt(sum((x - y)^2))}
+#' is used.
+#'
+#' @param base_optimizer (`Optimizer`)\cr
 #' An \code{Optimizer} object, which can be created via
-#' \code{\link[optimizeR]{Optimizer}}.
-#' It numerically solves the optimization problems in the partitions.
-#' @param iterations
-#' An \code{integer}, the maximum number of iterations through
-#' \code{partitions} before the alternating optimization process is terminated.
-#' Can also be \code{Inf}, in which case \code{tolerance} is responsible for the
-#' termination.
-#' The default is \code{10}.
-#' @param tolerance
-#' A non-negative \code{numeric}. The alternating optimization terminates
-#' prematurely (i.e., before \code{interations} is reached) if the euclidean
-#' distance between the current estimate and the one from the last iteration is
-#' smaller than \code{tolerance}.
-#' The default is \code{1e-6}.
-#' @param f_partition
-#' A \code{list} of the same length as \code{partition}.
-#' The \code{i}-th element can be a \code{function} that computes the value of
-#' \code{f} for the \code{i}-th parameter set defined in \code{partition}.
-#' The \code{function} must be of the form
-#' \code{function(theta_part, theta_rest, ...)}, where
-#' - \code{theta_part} receives the parameter set for the current partition
-#'   (this argument can be named differently),
-#' - \code{theta_rest} receives the remaining parameters
-#'   (this argument must be named \code{theta_rest}),
-#' - \code{...} receives the additional arguments to \code{f}.
-#' Alternatively, it can be \code{NULL}, in which case \code{f} is used.
-#' @param joint_end
-#' If \code{TRUE}, the parameter set is optimized jointly after the alternating
-#' optimization process is terminated.
-#' The default is \code{FALSE}.
-#' @param verbose
-#' If \code{TRUE}, full tracing details are printed during the alternating
-#' optimization process.
-#' The default is \code{FALSE}.
+#' \code{\link[optimizeR]{Optimizer}}. It numerically solves the sub-problems.
+#'
+#' By default, the \code{\link[stats]{optim}} optimizer is used. If another
+#' optimizer is specified, the arguments \code{gradient}, \code{lower}, and
+#' \code{upper} are ignored.
+#'
+#' @param verbose (`logical(1)`)\cr
+#' Whether to print tracing details during the alternating optimization
+#' process.
 #'
 #' @return
-#' A \code{list} with the elements
-#' * \code{estimate}, the optimal parameter vector found,
-#' * \code{value}, the value of \code{f} at \code{estimate},
-#' * \code{sequence}, a \code{data.frame} of the function values, estimates and
-#'   computation times in the single iterations and partitions,
-#' * and \code{seconds}, the overall computation time in seconds.
+#' A \code{list} with the following elements:
+#' * \code{estimate} is the parameter vector at termination.
+#' * \code{value} is the function value at termination.
+#' * \code{details} is a `data.frame` with full information about the procedure:
+#'   For each iteration (column `iteration`) it contains the function value
+#'   (column `value`), parameter values (columns starting with `p` followed by
+#'   the parameter index), the active parameter block (columns starting with `b`
+#'   followed by the parameter index, where `1` stands for a parameter contained
+#'   in the active parameter block and `0` if not), computation times in seconds
+#'   (column `seconds`), and a code that summarizes whether the update got
+#'   accepted (column `update_code`, where `0` stands for an accepted update,
+#'   `1` for a rejected update due to an error when solving the sub-problem, and
+#'   `2` for a rejected update because it did not improve the function value).
+#' * \code{seconds} is the overall computation time in seconds.
+#' * \code{stopping_reason} is a message why the procedure has terminated.
 #'
 #' @examples
-#' # definition of the Himmelblau function
-#' himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
+#' # Example 1: Minimization of Himmelblau's function --------------------------
 #'
-#' # alternating minimization separately for x_1 and x_2
-#' # parameter restriction: -5 <= x_1, x_2 <= 5
+#' himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
+#' ao(f = himmelblau, initial = c(0, 0))
+#'
+#' # Example 2: Maximization of 2-class Gaussian mixture log-likelihood --------
+#'
+#' # target arguments:
+#' # - class means mu (2, unrestricted)
+#' # - class standard deviations sd (2, must be non-negative)
+#' # - class proportion lambda (only 1 for identification, must be in [0, 1])
+#'
+#' normal_mixture_llk <- function(mu, sd, lambda, data) {
+#'   c1 <- lambda * dnorm(data, mu[1], sd[1])
+#'   c2 <- (1 - lambda) * dnorm(data, mu[2], sd[2])
+#'   sum(log(c1 + c2))
+#' }
+#'
 #' ao(
-#'   f = himmelblau, p = c(0, 0), partition = list(1, 2), iterations = Inf,
-#'   base_optimizer = optimizeR::Optimizer$new(
-#'     which = "stats::optim", lower = -5, upper = 5, method = "L-BFGS-B"
-#'   )
+#'   f = normal_mixture_llk,
+#'   initial = c(2, 4, 1, 1, 0.5),
+#'   target = c("mu", "sd", "lambda"),
+#'   npar = c(2, 2, 1),
+#'   data = datasets::faithful$eruptions,
+#'   partition = "random",
+#'   minimize = FALSE,
+#'   lower = c(-Inf, -Inf, 0, 0, 0),
+#'   upper = c(Inf, Inf, Inf, Inf, 1)
 #' )
 #'
 #' @export
 
 ao <- function(
-    f, p, ..., partition = as.list(1:length(p)),
-    base_optimizer = optimizeR::Optimizer$new("stats::optim"),
-    iterations = 10, tolerance = 1e-6,
-    f_partition = vector(mode = "list", length = length(partition)),
-    joint_end = FALSE, verbose = FALSE) {
-  ### input checks
-  if (missing(f)) {
-    cli::cli_abort(
-      "Please specify {.var f}",
-      call = NULL
-    )
-  } else if (!checkmate::test_function(f)) {
-    cli::cli_abort(
-      "{.var f} must be a function",
-      call = NULL
-    )
-  }
-  if (missing(p)) {
-    cli::cli_abort(
-      "Please specify {.var p}",
-      call = NULL
-    )
-  } else if (!oeli::test_numeric_vector(p)) {
-    cli::cli_abort(
-      "{.var p} must be a vector of initial values",
-      call = NULL
-    )
-  }
-  if (!checkmate::test_list(partition)) {
-    cli::cli_abort(
-      "{.var partition} must be a {.cls list}",
-      call = NULL
-    )
-  } else if (!setequal(unlist(partition), seq_along(p)) ||
-    any(sapply(partition, length) == 0)) {
-    cli::cli_abort(
-      "{.var partition} must only contain vectors of indices of {.var p}",
-      call = NULL
-    )
-  }
-  if (!inherits(base_optimizer, "Optimizer")) {
-    cli::cli_abort(
-      c(
-        "{.var base_optimizer} must be an {.cls Optimizer} object",
-        "i" = "Use {.fun optimizeR::Optimizer} to create such an object"
-      ),
-      call = NULL
-    )
-  }
-  if (!checkmate::test_number(iterations, lower = 1, finite = FALSE)) {
-    cli::cli_abort(
-      "{.var iterations} must be an integer greater or equal 1",
-      call = NULL
-    )
-  } else if (is.finite(iterations)) {
-    iterations <- as.integer(iterations)
-  }
-  if (!checkmate::test_number(tolerance, lower = 0, finite = TRUE)) {
-    cli::cli_abort(
-      "{.var tolerance} must be a single, non-negative number",
-      call = NULL
-    )
-  } else if (tolerance == 0 && identical(iterations, Inf)) {
-    cli::cli_abort(
-      "{.var tolerance} cannot be 0 if {.var iterations} is Inf",
-      call = NULL
-    )
-  }
-  if (!checkmate::test_list(f_partition)) {
-    cli::cli_abort(
-      "{.var f_partition} must be a {.cls list}",
-      call = NULL
-    )
-  } else if (length(f_partition) != length(partition)) {
-    cli::cli_abort(
-      "{.var f_partition} must have the same length as {.var partition}",
-      call. = FALSE
-    )
-  }
-  if (!checkmate::test_flag(joint_end)) {
-    cli::cli_abort(
-      "{.var joint_end} must be TRUE or FALSE",
-      call = NULL
-    )
-  }
-  if (!checkmate::test_flag(verbose)) {
-    cli::cli_abort(
-      "{.var verbose} must be TRUE or FALSE",
-      call = NULL
-    )
-  }
-
-  ### preparing partitions
-  npar <- length(p)
-  for (part in seq_along(partition)) {
-    if (!is.function(f_partition[[part]])) {
-      f_partition[[part]] <- function(theta_part, theta_rest, ...) {
-        p_ind <- partition[[part]]
-        theta <- numeric(npar)
-        theta[p_ind] <- theta_part
-        theta[-p_ind] <- theta_rest
-        out <- f(theta, ...)
-        if ("gradient" %in% names(attributes(out))) {
-          gradient <- attr(out, "gradient")
-          if (is.numeric(gradient) && is.vector(gradient)) {
-            attr(out, "gradient") <- gradient[p_ind]
-          }
-        }
-        if ("hessian" %in% names(attributes(out))) {
-          hessian <- attr(out, "hessian")
-          if (is.numeric(hessian) && is.matrix(hessian)) {
-            attr(out, "hessian") <- hessian[p_ind, p_ind, drop = FALSE]
-          }
-        }
-        out
-      }
-    } else {
-      add_arguments_exist <- length(list(...)) > 0
-      if (length(formals(f_partition[[part]])) < 2 + add_arguments_exist) {
-        cli::cli_abort(
-          "{.var f_partition[[{part}]]} must have two arguments
-          {ifelse(add_arguments_exist, 'and the ... argument', '')}",
-          call = NULL
-        )
-      }
-      if (!identical(names(formals(f_partition[[part]]))[2], "theta_rest")) {
-        cli::cli_abort(
-          "{.var f_partition[[{part}]]} must have a second argument named
-          {.var theta_rest}",
-          call = NULL
-        )
-      }
-    }
-  }
-
-  ### alternating optimization
-  exit_flag <- FALSE
-  est <- p
-  f_initial <- try(f(est, ...), silent = TRUE)
-  if (!checkmate::test_number(f_initial, finite = TRUE)) {
-    cli::cli_abort(
-      "{.var f(p)} is not a single, finite number",
-      call = NULL
-    )
-  }
-  seq <- structure(
-    data.frame(t(c(0, NA_integer_, f_initial, 0, est))),
-    names = c("iteration", "partition", "value", "seconds", paste0("p", 1:npar))
+    f,
+    initial,
+    target = NULL,
+    npar = NULL,
+    gradient = NULL,
+    ...,
+    partition = "sequential",
+    new_block_probability = 0.3,
+    minimum_block_number = 2,
+    minimize = TRUE,
+    lower = -Inf,
+    upper = Inf,
+    iteration_limit = Inf,
+    seconds_limit = Inf,
+    tolerance_value = 1e-6,
+    tolerance_parameter = 1e-6,
+    tolerance_parameter_norm = function(x, y) sqrt(sum((x - y)^2)),
+    base_optimizer = Optimizer$new("stats::optim", method = "L-BFGS-B"),
+    verbose = FALSE) {
+  ### input checks and building of objects
+  ao_input_check(
+    "initial",
+    oeli::check_numeric_vector(initial, any.missing = FALSE)
   )
-  iteration <- 1
-  while (iteration <= iterations) {
-    if (exit_flag) {
-      break
+  if (is.null(npar)) {
+    npar <- length(initial)
+  }
+  objective <- Objective$new(f = f, target = target, npar = npar, ...)
+  npar <- sum(objective$npar)
+  ao_input_check(
+    "initial",
+    oeli::check_numeric_vector(initial, len = npar)
+  )
+  ao_input_check(
+    "lower",
+    oeli::check_numeric_vector(lower, any.missing = FALSE)
+  )
+  if (length(lower) == 1) {
+    lower <- rep(lower, npar)
+  }
+  ao_input_check(
+    "lower",
+    oeli::check_numeric_vector(lower, len = npar)
+  )
+  ao_input_check(
+    "initial",
+    if (any(initial < lower)) "Must respect lower limit {.var lower}" else TRUE
+  )
+  ao_input_check(
+    "upper",
+    oeli::check_numeric_vector(upper, any.missing = FALSE)
+  )
+  if (length(upper) == 1) {
+    upper <- rep(upper, npar)
+  }
+  ao_input_check(
+    "upper",
+    oeli::check_numeric_vector(upper, len = npar)
+  )
+  ao_input_check(
+    "initial",
+    if (any(initial > upper)) "Must respect upper limit {.var upper}" else TRUE
+  )
+  if (!is.null(gradient)) {
+    gradient <- Objective$new(f = gradient, target = target, npar = npar, ...)
+  }
+  ao_input_check(
+    "base_optimizer",
+    checkmate::check_class(base_optimizer, "Optimizer")
+  )
+  if (base_optimizer$label != "stats::optim") {
+    cli::cli_warn(
+      "Arguments {.var gradient}, {.var lower}, and {.var upper} are ignored"
+    )
+  }
+  procedure <- Procedure$new(
+    npar = npar,
+    partition = partition,
+    new_block_probability = new_block_probability,
+    minimum_block_number = minimum_block_number,
+    verbose = verbose,
+    minimize = minimize,
+    iteration_limit = iteration_limit,
+    seconds_limit = seconds_limit,
+    tolerance_value = tolerance_value,
+    tolerance_parameter = tolerance_parameter,
+    tolerance_parameter_norm = tolerance_parameter_norm
+  )
+
+  ### build sub-problem template
+  solve_sub_problem <- function(parameter_block, parameter_fixed, block) {
+    ### build block objective function
+    block_objective <- function(parameter_block) {
+      theta <- numeric(npar)
+      theta[block] <- parameter_block
+      theta[-block] <- parameter_fixed
+      objective$evaluate(theta)
     }
-    if (verbose) {
-      cat("iteration", iteration, "of", iterations, "\n")
-    }
-    for (part in seq_along(partition)) {
-      if (verbose) {
-        cat("- partition", part, "of", length(partition), ": ")
-      }
-      p_ind <- partition[[part]]
-      f_part_out <- base_optimizer$minimize(
-        objective = f_partition[[part]], initial = est[p_ind],
-        theta_rest = est[-p_ind], ...
-      )
-      if (isTRUE(f_part_out$error)) {
-        cli::cli_abort(
-          f_part_out$error_message,
-          call = NULL
-        )
-      }
-      est[p_ind] <- f_part_out[["parameter"]]
-      value <- f_part_out[["value"]]
-      if (verbose) {
-        cat("f =", value, "\n")
-      }
-      seq <- rbind(seq, c(iteration, part, value, f_part_out[["seconds"]], est))
-      if (nrow(seq) > length(partition)) {
-        curr <- as.numeric(seq[nrow(seq) - length(partition), -(1:4)])
-        last <- as.numeric(seq[nrow(seq), -(1:4)])
-        dist <- sqrt(sum(curr - last)^2)
-        if (dist < tolerance) {
-          exit_flag <- TRUE
-          if (verbose) {
-            cat("tolerance reached : distance =", dist, "<", tolerance, "\n")
-          }
-          break
+
+    ### build block gradient function and set arguments for 'stats::optim'
+    if (base_optimizer$label == "stats::optim") {
+      block_gradient <- if (is.null(gradient)) {
+        NULL
+      } else {
+        function(parameter_block, ...) {
+          theta <- numeric(npar)
+          theta[block] <- parameter_block
+          theta[-block] <- parameter_fixed
+          gradient$evaluate(theta)[block]
         }
       }
+      base_optimizer$set_arguments(
+        "gr" = block_gradient, "lower" = lower[block], "upper" = upper[block]
+      )
     }
-    iteration <- iteration + 1
+
+    ### solve sub-problem
+    base_optimizer$optimize(
+      objective = block_objective,
+      initial = parameter_block,
+      direction = ifelse(procedure$minimize, "min", "max")
+    )
   }
-  if (joint_end) {
-    if (verbose) {
-      cat("joint optimization in the end : ")
+
+  ### start alternating optimization
+  procedure$initialize_details(
+    initial_parameter = initial,
+    initial_value = objective$evaluate(initial)
+  )
+  while (TRUE) {
+    ### check stopping criteria
+    if (procedure$check_stopping()) {
+      break
+    } else {
+      procedure$iteration <- procedure$iteration + 1L
     }
-    f_joint_out <- base_optimizer$minimize(
-      objective = f, initial = est, ...
-    )
-    est <- f_joint_out[["parameter"]]
-    value <- f_joint_out[["value"]]
-    seq <- rbind(
-      seq, c(NA_integer_, NA_integer_, value, f_joint_out[["seconds"]], est)
-    )
-    if (verbose) {
-      cat("f =", value, "\n")
+
+    ### optimize over each parameter block in partition
+    for (block in procedure$get_partition()) {
+      procedure$block <- block
+
+      ### optimize block objective function
+      sub_problem_out <- solve_sub_problem(
+        parameter_block = procedure$get_parameter_latest("block"),
+        parameter_fixed = procedure$get_parameter_latest("fixed"),
+        block = block
+      )
+
+      ### check acceptance and update
+      procedure$update_details(
+        value = sub_problem_out[["value"]],
+        parameter_block = sub_problem_out[["parameter"]],
+        seconds = sub_problem_out[["seconds"]],
+        error = sub_problem_out[["error"]]
+      )
     }
   }
 
   ### return results
-  list(
-    "value" = value,
-    "estimate" = est,
-    "sequence" = seq,
-    "seconds" = sum(seq$seconds)
-  )
+  procedure$output
 }